Allocations: rename Budgets to Overview
* Move Header inside
* Move Allocations list inside

These are not things we would want to appear on a BudgetDetail page
