+++ apps/allocations/app/components/App/App.js
@@ -1,42 +1,21 @@
 import React from 'react'
 import { useAragonApi } from '../../api-react'
-import { Button, Header, IconPlus, Main, SidePanel, SyncIndicator } from '@aragon/ui'
+import { Main, SidePanel, SyncIndicator } from '@aragon/ui'
 
 import { IdentityProvider } from '../LocalIdentityBadge/IdentityManager'
-import { Empty } from '../Card'
-import { NewBudget } from '../Panel'
-import { AllocationsHistory, Budgets } from '.'
+import { Overview } from '.'
 import { usePanel } from '../../context/Panel'
 
 const App = () => {
   const { api, appState } = useAragonApi()
-  const { allocations = [], budgets = [], isSyncing = true } = appState
+  const { isSyncing = true } = appState
 
   const { panel, panelOpen, setPanel } = usePanel()
-
-  const saveBudget = ({ amount, name, token }) => {
-    api
-      .newAccount(
-        name,             // _metadata
-        token.address,    // _token
-        true,             // hasBudget
-        amount
-      )
-      .toPromise()
-    setPanel(null)
-  }
 
   // TODO: Fix this
   // eslint-disable-next-line
   const onExecutePayout = (accountId, payoutId) => {
     api.runPayout(accountId, payoutId).toPromise()
-  }
-
-  const onNewBudget = () => {
-    setPanel({
-      content: NewBudget,
-      data: { heading: 'New budget', saveBudget },
-    })
   }
 
   const handleResolveLocalIdentity = address =>
@@ -53,27 +32,8 @@ const App = () => {
         onResolve={handleResolveLocalIdentity}
         onShowLocalIdentityModal={handleShowLocalIdentityModal}
       >
-        {budgets.length === 0
-          ? <Empty action={onNewBudget} isSyncing={isSyncing} />
-          : (
-            <React.Fragment>
-              <Header
-                primary="Allocations"
-                secondary={
-                  <Button
-                    mode="strong"
-                    icon={<IconPlus />}
-                    onClick={onNewBudget}
-                    label="New budget"
-                  />
-                }
-              />
-              <Budgets />
-              <SyncIndicator visible={isSyncing} />
-            </React.Fragment>
-          )
-        }
-        { !!allocations.length && <AllocationsHistory allocations={allocations} /> }
+        <Overview />
+        <SyncIndicator visible={isSyncing} />
         <SidePanel
           title={(panel && panel.data.heading) || ''}
           opened={panelOpen}
