+++ apps/allocations/app/components/App/index.js
@@ -1,2 +1,2 @@
-export { default as Budgets } from './Budgets'
+export { default as Overview } from './Overview'
 export { default as AllocationsHistory } from './AllocationsHistory'
